Risolto tutti i problemi di esecuzione e di visualizzazione delle matrici, risolto il problema del compilare con gcc anziche con g++, gli eseguibili sono tornati ad essere quelli di sempre: ./elanja e ./test, aggiornato il file README.

+++ h/util.h
@@ -1,2 +1,2 @@
 /* Interaction function */
-extern void intraction(double *p, double *q, double *A, double *R, int *degree, int *A_degree, int *B_degree);
+extern void intraction(int *p, int *q, int *A, int *R, int *degree, int *A_degree, int *B_degree);
